requirements.txt
bcrypt
fastapi[standard]
mysql-connector-python
pydantic
python-dotenv
requests